# core/vm: optimize the mstore opcode with loop unrolling